refactor: put stuff in separate files
