+++ lib/socket_connector.dart
@@ -1,3 +1,4 @@
 library socket_connector.dart;
 
 export 'package:socket_connector/src/socket_connector.dart';
+export 'package:socket_connector/src/types.dart';
